requirements.txt
flask==3.1.1
flask-restx==1.3.0
psycopg2-binary==2.9.10
requests==2.32.3
python-dotenv==1.1.0